Revise fix to what works
