+++ src/client.js
@@ -1,1 +1,21 @@
-import './client';
+/* eslint-disable react/jsx-filename-extension  */
+import React from 'react';
+import { hydrate } from 'react-dom';
+import { BrowserRouter } from 'react-router-dom';
+import { ensureReady, After } from '@jaredpalmer/after';
+import routes from './app/routes';
+
+const root = document.getElementById('root');
+
+ensureReady(routes).then(data =>
+  hydrate(
+    <BrowserRouter>
+      <After data={data} routes={routes} />
+    </BrowserRouter>,
+    root,
+  ),
+);
+
+if (module.hot) {
+  module.hot.accept();
+}
