--- src/client/index.jsx
@@ -1,30 +1,0 @@
-import React from 'react';
-import { hydrate } from 'react-dom';
-import { BrowserRouter } from 'react-router-dom';
-import { ensureReady, After } from '@jaredpalmer/after';
-import routes from '../app/routes';
-
-const root = document.getElementById('root');
-
-ensureReady(routes).then(data =>
-  hydrate(
-    <BrowserRouter>
-      <After data={data} routes={routes} />
-    </BrowserRouter>,
-    root,
-  ),
-);
-
-if (module.hot) {
-  module.hot.accept('../app/routes', () => {
-    const routes = require('../app/routes'.default);
-    ensureReady(routes).then(data =>
-      hydrate(
-        <BrowserRouter>
-          <After data={data} routes={routes} />
-        </BrowserRouter>,
-        root,
-      ),
-    );
-  }
-}
